requirements.txt
SQLAlchemy~=2.0.19
alembic~=1.11.2
python-dotenv~=1.0.0
aiogram~=3.0.0rc1